+++ include/H2DE/H2DE.h
@@ -7,8 +7,12 @@
  * Hydren 2 Dimensional Engine
  * https://github.com/Hydrened/H2DE
  * 
- * Version: b-3.0.3 (Beta)
+ * Version: b-3.0.4 (Beta)
  * 
  * Copyright (C) 2024-2025
  */
 #include <H2DE_engine.h>
+#include <H2DE/H2DE_bar_object.h>
+#include <H2DE/H2DE_basic_object.h>
+#include <H2DE/H2DE_button_object.h>
+#include <H2DE/H2DE_text_object.h>