+++ include/H2DE/H2DE_camera.h
@@ -9,7 +9,7 @@ private:
     H2DE_CameraData data;
     H2DE_LevelPos pos;
 
-    H2DE_LevelObject* reference = nullptr;
+    H2DE_Object* reference = nullptr;
     bool lockedToReference = false;
     H2DE_LevelPadding lockedPadding = { 0.0f, 0.0f, 0.0f, 0.0f };
 
@@ -17,15 +17,16 @@ public:
     H2DE_Camera(H2DE_Engine* engine, const H2DE_CameraData& data);
     ~H2DE_Camera();
 
-    friend H2DE_LevelPos H2DE_GetCameraPos(H2DE_Engine* engine);
-    friend H2DE_LevelSize H2DE_GetCameraSize(H2DE_Engine* engine);
-    friend bool H2DE_CameraContains(H2DE_Engine* engine, const H2DE_Hitbox& hitbox);
-    friend void H2DE_SetCameraPos(H2DE_Engine* engine, const H2DE_LevelPos& pos);
-    friend void H2DE_SetCameraWidth(H2DE_Engine* engine, const float& width);
-    friend void H2DE_SetCameraSmoothing(H2DE_Engine* engine, const float& smoothing);
-    friend void H2DE_SetCameraReference(H2DE_Engine* engine, H2DE_LevelObject* object);
-    friend void H2DE_SetCameraLockedToReference(H2DE_Engine* engine, const bool& state);
-    friend void H2DE_SetCameraPadding(H2DE_Engine* engine, const H2DE_LevelPadding& padding);
+    friend H2DE_LevelPos H2DE_GetCameraPos(const H2DE_Engine* engine);
+    friend H2DE_LevelSize H2DE_GetCameraSize(const H2DE_Engine* engine);
+    friend bool H2DE_CameraContainsObject(const H2DE_Engine* engine, H2DE_Object* object);
+    friend bool H2DE_CameraContainsHitbox(const H2DE_Engine* engine, const H2DE_Hitbox& hitbox, bool absolute);
+    friend void H2DE_SetCameraPos(const H2DE_Engine* engine, const H2DE_LevelPos& pos);
+    friend void H2DE_SetCameraWidth(const H2DE_Engine* engine, float width);
+    friend void H2DE_SetCameraSmoothing(const H2DE_Engine* engine, float smoothing);
+    friend void H2DE_SetCameraReference(const H2DE_Engine* engine, H2DE_Object* object);
+    friend void H2DE_SetCameraLockedToReference(const H2DE_Engine* engine, bool state);
+    friend void H2DE_SetCameraPadding(const H2DE_Engine* engine, const H2DE_LevelPadding& padding);
 };
 
 #endif