+++ src/objects/H2DE_button_object.cpp
@@ -1,0 +1,22 @@
+#include "H2DE/H2DE_button_object.h"
+
+// INIT
+H2DE_ButtonObject::H2DE_ButtonObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_ButtonObjectData b) : H2DE_Object(engine, od), bod(b) {
+
+}
+
+H2DE_ButtonObject* H2DE_CreateButtonObject(H2DE_Engine* engine, const H2DE_ObjectData& od, const H2DE_ButtonObjectData& bod) {
+    H2DE_ButtonObject* object = new H2DE_ButtonObject(engine, od, bod);
+    engine->addObject(object);
+    return object;
+}
+
+// CLEANUP
+H2DE_ButtonObject::~H2DE_ButtonObject() {
+
+}
+
+// UPDATE
+void H2DE_ButtonObject::update() {
+    
+}